@@ -1,5 +1,15 @@
 module github.com/salrashid123/oauth2oidc
 
-go 1.15
+go 1.20
 
-require golang.org/x/oauth2 v0.0.0-20200902213428-5d25da1a8d43
+require (
+	github.com/golang-jwt/jwt v3.2.2+incompatible
+	golang.org/x/oauth2 v0.0.0-20200902213428-5d25da1a8d43
+)
+
+require (
+	github.com/golang/protobuf v1.4.2 // indirect
+	golang.org/x/net v0.0.0-20200822124328-c89045814202 // indirect
+	google.golang.org/appengine v1.6.6 // indirect
+	google.golang.org/protobuf v1.25.0 // indirect
+)
